Fix for copy of compound file >2Go.
Alain Barbet <alian@amisw.com>

+++ src/core/index/CompoundFileWriter.cpp
@@ -27,60 +27,60 @@ namespace Lucene
         entries = Collection<FileEntry>::newInstance();
         merged = false;
     }
-    
+
     CompoundFileWriter::~CompoundFileWriter()
     {
     }
-    
+
     DirectoryPtr CompoundFileWriter::getDirectory()
     {
         return DirectoryPtr(_directory);
     }
-    
+
     String CompoundFileWriter::getName()
     {
         return fileName;
     }
-    
+
     void CompoundFileWriter::addFile(const String& file)
     {
         if (merged)
             boost::throw_exception(IllegalStateException(L"Can't add extensions after merge has been called"));
-        
+
         if (file.empty())
             boost::throw_exception(IllegalArgumentException(L"file cannot be empty"));
-        
+
         if (!ids.add(file))
             boost::throw_exception(IllegalArgumentException(L"File " + file + L" already added"));
-        
+
         FileEntry entry;
         entry.file = file;
         entries.add(entry);
     }
-    
+
     void CompoundFileWriter::close()
     {
         if (merged)
             boost::throw_exception(IllegalStateException(L"Merge already performed"));
-        
+
         if (entries.empty())
             boost::throw_exception(IllegalStateException(L"No entries to merge have been defined"));
-        
+
         merged = true;
-        
+
         DirectoryPtr directory(_directory);
-        
+
         // open the compound stream
         IndexOutputPtr os;
         LuceneException finally;
         try
         {
             os = directory->createOutput(fileName);
-            
+
             // Write the number of entries
             os->writeVInt(entries.size());
-            
-            // Write the directory with all offsets at 0. Remember the positions of directory entries so that we 
+
+            // Write the directory with all offsets at 0. Remember the positions of directory entries so that we
             // can adjust the offsets later
             int64_t totalSize = 0;
             for (Collection<FileEntry>::iterator fe = entries.begin(); fe != entries.end(); ++fe)
@@ -90,13 +90,13 @@ namespace Lucene
                 os->writeString(fe->file);
                 totalSize += directory->fileLength(fe->file);
             }
-            
-            // Pre-allocate size of file as optimization - this can potentially help IO performance as we write the 
-            // file and also later during searching.  It also uncovers a disk-full situation earlier and hopefully 
+
+            // Pre-allocate size of file as optimization - this can potentially help IO performance as we write the
+            // file and also later during searching.  It also uncovers a disk-full situation earlier and hopefully
             // without actually filling disk to 100%
             int64_t finalLength = totalSize + os->getFilePointer();
             os->setLength(finalLength);
-            
+
             // Open the files and copy their data into the stream. Remember the locations of each file's data section.
             ByteArray buffer(ByteArray::newInstance(16384));
             for (Collection<FileEntry>::iterator fe = entries.begin(); fe != entries.end(); ++fe)
@@ -104,17 +104,17 @@ namespace Lucene
                 fe->dataOffset = os->getFilePointer();
                 copyFile(*fe, os, buffer);
             }
-            
+
             // Write the data offsets into the directory of the compound stream
             for (Collection<FileEntry>::iterator fe = entries.begin(); fe != entries.end(); ++fe)
             {
                 os->seek(fe->directoryOffset);
                 os->writeLong(fe->dataOffset);
             }
-            
+
             BOOST_ASSERT(finalLength == os->length());
-            
-            // Close the output stream. Set the os to null before trying to close so that if an exception occurs during 
+
+            // Close the output stream. Set the os to null before trying to close so that if an exception occurs during
             // the close, the finally clause below will not attempt to close the stream the second time.
             IndexOutputPtr tmp(os);
             os.reset();
@@ -124,10 +124,10 @@ namespace Lucene
         {
             finally = e;
         }
-        
+
         if (os)
         {
-            try 
+            try
             {
                 os->close();
             }
@@ -137,7 +137,7 @@ namespace Lucene
         }
         finally.throwException();
     }
-    
+
     void CompoundFileWriter::copyFile(const FileEntry& source, IndexOutputPtr os, ByteArray buffer)
     {
         IndexInputPtr is;
@@ -146,15 +146,15 @@ namespace Lucene
         try
         {
             int64_t startPtr = os->getFilePointer();
-            
+
             is = directory->openInput(source.file);
             int64_t length = is->length();
             int64_t remainder = length;
-            int32_t chunk = buffer.size();
-            
+            int64_t chunk = buffer.size();
+
             while (remainder > 0)
             {
-                int32_t len = std::min(chunk, (int32_t)remainder);
+                int32_t len = (int32_t)std::min(chunk, remainder);
                 is->readBytes(buffer.get(), 0, len, false);
                 os->writeBytes(buffer.get(), len);
                 remainder -= len;
@@ -164,15 +164,15 @@ namespace Lucene
                     checkAbort->work(80);
                 }
             }
-            
+
             // Verify that remainder is 0
             if (remainder != 0)
             {
-                boost::throw_exception(IOException(L"Non-zero remainder length after copying: " + StringUtils::toString(remainder) + 
+                boost::throw_exception(IOException(L"Non-zero remainder length after copying: " + StringUtils::toString(remainder) +
                                                    L" (id: " + source.file + L", length: " + StringUtils::toString(length) +
                                                    L", buffer size: " + StringUtils::toString(chunk) + L")"));
             }
-            
+
             // Verify that the output length diff is equal to original file
             int64_t endPtr = os->getFilePointer();
             int64_t diff = endPtr - startPtr;
@@ -186,7 +186,7 @@ namespace Lucene
         {
             finally = e;
         }
-        
+
         if (is)
             is->close();
         finally.throwException();
